Added Last Messages Read
Added Last Messages Read ability.
Bob read 1 message from forum 1.
Joe Cool has read no messages at all.
Also, fixed jUnit tests.

+++ src/test/java/com/bbs/services/AuthoritiesServiceTests.java
@@ -2,6 +2,8 @@ package com.bbs.services;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import java.math.BigInteger;
+
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -23,6 +25,6 @@ public class AuthoritiesServiceTests {
 	public void testSave() {
 		Authority authority = new Authority("fred", "ROLE_USER");
 		authority=service.save(authority);
-		assertEquals(3, authority.getId());
+		assertEquals(BigInteger.valueOf(3), authority.getId());
 	}
 }
