+++ src/test/java/com/bbs/services/UsersServiceTests.java
@@ -2,6 +2,8 @@ package com.bbs.services;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import java.math.BigInteger;
+
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -28,6 +30,6 @@ public class UsersServiceTests {
 		User user = new User("fred", passwordEncoder.encode("Pass123"), true);
 		System.out.println(passwordEncoder.encode("Joe Cool"));
 		user=service.save(user);
-		assertEquals(3, user.getId());
+		assertEquals(BigInteger.valueOf(3), user.getId());
 	}
 }
